hatsumei/backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask_migrate
python-dotenv
flask-cors
flask-marshmallow
marshmallow-sqlalchemy